backend/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.32.0
pydantic==2.9.2
python-dotenv==1.0.1
python-multipart==0.0.12
openai==1.55.3
dashscope>=1.22.0
